requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
opencv-python>=4.5.0
Pillow>=9.0.0
tqdm>=4.64.0
pathlib2>=2.3.0
scikit-learn>=1.3.0
timm>=0.9.0
scenedetect>=0.6.0